Adding causal lm test cases
