+++ tests/conftest.py
@@ -5,9 +5,6 @@
 #
 # -----------------------------------------------------------------------------
 
-import os
-import shutil
-from pathlib import Path
 
 import pytest
 from transformers import logging
@@ -118,25 +115,26 @@ def qeff_models_clean_up(qeff_dir=QEFF_HOME):
         qeff_dir: Can be a string (file/dir path), PosixPath, or list of strings/PosixPath objects
                  If a file path is provided, its parent directory will be deleted
     """
-    if isinstance(qeff_dir, (str, Path)):
-        paths = [qeff_dir]
-    else:
-        paths = qeff_dir
+    pass
+    # if isinstance(qeff_dir, (str, Path)):
+    #     paths = [qeff_dir]
+    # else:
+    #     paths = qeff_dir
 
-    for path in paths:
-        try:
-            path_str = str(path)
-            if os.path.isfile(path_str):
-                dir_to_delete = os.path.dirname(path_str)
-                if os.path.exists(dir_to_delete):
-                    shutil.rmtree(dir_to_delete)
-                    print(f"\n.............Cleaned up {dir_to_delete}")
-            elif os.path.isdir(path_str):
-                if os.path.exists(path_str):
-                    shutil.rmtree(path_str)
-                    print(f"\n.............Cleaned up {path_str}")
-        except Exception as e:
-            print(f"\n.............Error cleaning up {path}: {e}")
+    # for path in paths:
+    #     try:
+    #         path_str = str(path)
+    #         if os.path.isfile(path_str):
+    #             dir_to_delete = os.path.dirname(path_str)
+    #             if os.path.exists(dir_to_delete):
+    #                 shutil.rmtree(dir_to_delete)
+    #                 print(f"\n.............Cleaned up {dir_to_delete}")
+    #         elif os.path.isdir(path_str):
+    #             if os.path.exists(path_str):
+    #                 shutil.rmtree(path_str)
+    #                 print(f"\n.............Cleaned up {path_str}")
+    #     except Exception as e:
+    #         print(f"\n.............Error cleaning up {path}: {e}")
 
 
 @pytest.fixture
