--- tests/transformers/models/causal_lm_models/test_causal_lm_blocking_hqkv.py
@@ -1,379 +1,0 @@
-# -----------------------------------------------------------------------------
-#
-# Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
-# SPDX-License-Identifier: BSD-3-Clause
-#
-# -----------------------------------------------------------------------------
-
-import json
-import os
-
-import pytest
-from transformers import AutoConfig
-
-from QEfficient.utils.test_utils import ModelConfig
-
-from .check_causal_models import (
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100,
-    get_custom_n_layers,
-)
-
-CONFIG_PATH = os.path.join(os.path.dirname(__file__), "../../../configs/causal_model_configs.json")
-with open(CONFIG_PATH, "r") as f:
-    config_data = json.load(f)
-    blockedKV_models = config_data["blockedKV_causal_lm_models"]
-test_models_blockedKV = [model["model_name"] for model in blockedKV_models]
-model_config_dict = {model["model_name"]: model for model in blockedKV_models}
-
-
-@pytest.mark.full_layers
-@pytest.mark.llm_model
-@pytest.mark.on_qaic
-@pytest.mark.parametrize("model_name", test_models_blockedKV[:1])
-def test_full_causal_all_blocking_pytorch_vs_kv_vs_ort_vs_ai100(model_name, manual_cleanup):
-    HEAD_BLOCK_SIZE = 8
-    NUM_KV_BLOCKS = 2
-    NUM_Q_BLOCKS = 2
-
-    # head blocking only
-    qaic_config = dict(enable_blocking=True, head_block_size=HEAD_BLOCK_SIZE)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, manual_cleanup=manual_cleanup, num_devices=4
-    )
-
-    # kv blocking only
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, manual_cleanup=manual_cleanup, num_devices=4
-    )
-
-    # q block only
-    qaic_config = dict(enable_blocking=True, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, manual_cleanup=manual_cleanup, num_devices=4
-    )
-
-    # qkv blocking
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, manual_cleanup=manual_cleanup, num_devices=4
-    )
-
-    # head qkv blocking
-    qaic_config = dict(
-        enable_blocking=True,
-        head_block_size=HEAD_BLOCK_SIZE,
-        num_kv_blocks=NUM_KV_BLOCKS,
-        num_q_blocks=NUM_Q_BLOCKS,
-    )
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, manual_cleanup=manual_cleanup, num_devices=4
-    )
-
-
-@pytest.mark.few_layers
-@pytest.mark.llm_model
-@pytest.mark.on_qaic
-@pytest.mark.parametrize("model_name", test_models_blockedKV[:1])
-def test_few_causal_all_blocking_pytorch_vs_kv_vs_ort_vs_ai100(model_name, manual_cleanup):
-    HEAD_BLOCK_SIZE = 8
-    NUM_KV_BLOCKS = 2
-    NUM_Q_BLOCKS = 2
-    n_layer = get_custom_n_layers(model_name)
-    # head blocking only
-    qaic_config = dict(enable_blocking=True, head_block_size=HEAD_BLOCK_SIZE)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, manual_cleanup=manual_cleanup
-    )
-
-    # kv blocking only
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, manual_cleanup=manual_cleanup
-    )
-
-    # q block only
-    qaic_config = dict(enable_blocking=True, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, manual_cleanup=manual_cleanup
-    )
-
-    # qkv blocking
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, manual_cleanup=manual_cleanup
-    )
-
-    # head qkv blocking
-    qaic_config = dict(
-        enable_blocking=True,
-        head_block_size=HEAD_BLOCK_SIZE,
-        num_kv_blocks=NUM_KV_BLOCKS,
-        num_q_blocks=NUM_Q_BLOCKS,
-    )
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, manual_cleanup=manual_cleanup
-    )
-
-
-@pytest.mark.dummy_layers
-@pytest.mark.llm_model
-@pytest.mark.on_qaic
-@pytest.mark.parametrize("model_name", test_models_blockedKV[:1])
-def test_dummy_causal_all_blocking_pytorch_vs_kv_vs_ort_vs_ai100(model_name, manual_cleanup):
-    HEAD_BLOCK_SIZE = 8
-    NUM_KV_BLOCKS = 2
-    NUM_Q_BLOCKS = 2
-
-    hf_config = AutoConfig.from_pretrained(
-        model_name,
-        trust_remote_code=model_name in ModelConfig.EXTERNAL_MODELS,
-        **model_config_dict[model_name].get("additional_params", {}),
-    )
-    n_layer = -1
-    if model_name in ModelConfig.QUANTIZED_MODELS:
-        n_layer = get_custom_n_layers(model_name)
-        hf_config = None
-    # head blocking only
-    qaic_config = dict(enable_blocking=True, head_block_size=HEAD_BLOCK_SIZE)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, config=hf_config, manual_cleanup=manual_cleanup
-    )
-
-    # kv blocking only
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, config=hf_config, manual_cleanup=manual_cleanup
-    )
-
-    # q block only
-    qaic_config = dict(enable_blocking=True, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, config=hf_config, manual_cleanup=manual_cleanup
-    )
-
-    # qkv blocking
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, config=hf_config, manual_cleanup=manual_cleanup
-    )
-
-    # head qkv blocking
-    qaic_config = dict(
-        enable_blocking=True,
-        head_block_size=HEAD_BLOCK_SIZE,
-        num_kv_blocks=NUM_KV_BLOCKS,
-        num_q_blocks=NUM_Q_BLOCKS,
-    )
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name, qaic_config=qaic_config, n_layer=n_layer, config=hf_config, manual_cleanup=manual_cleanup
-    )
-
-
-@pytest.mark.full_layers
-@pytest.mark.llm_model
-@pytest.mark.on_qaic
-@pytest.mark.parametrize("model_name", test_models_blockedKV[:1])
-def test_full_causal_all_blocking_pytorch_vs_kv_vs_ort_vs_ai100_CB(model_name, manual_cleanup):
-    HEAD_BLOCK_SIZE = 8
-    NUM_KV_BLOCKS = 2
-    NUM_Q_BLOCKS = 2
-
-    # head blocking only
-    qaic_config = dict(enable_blocking=True, head_block_size=HEAD_BLOCK_SIZE)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-        num_devices=4,
-    )
-
-    # kv blocking only
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-        num_devices=4,
-    )
-
-    # q block only
-    qaic_config = dict(enable_blocking=True, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-        num_devices=4,
-    )
-
-    # qkv blocking
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-        num_devices=4,
-    )
-
-    # head qkv blocking
-    qaic_config = dict(
-        enable_blocking=True,
-        head_block_size=HEAD_BLOCK_SIZE,
-        num_kv_blocks=NUM_KV_BLOCKS,
-        num_q_blocks=NUM_Q_BLOCKS,
-    )
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-        num_devices=4,
-    )
-
-
-@pytest.mark.few_layers
-@pytest.mark.llm_model
-@pytest.mark.on_qaic
-@pytest.mark.parametrize("model_name", test_models_blockedKV[:1])
-def test_few_causal_all_blocking_pytorch_vs_kv_vs_ort_vs_ai100_CB(model_name, manual_cleanup):
-    HEAD_BLOCK_SIZE = 8
-    NUM_KV_BLOCKS = 2
-    NUM_Q_BLOCKS = 2
-    n_layer = get_custom_n_layers(model_name)
-    # head blocking only
-    qaic_config = dict(enable_blocking=True, head_block_size=HEAD_BLOCK_SIZE)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-    # kv blocking only
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-    # q block only
-    qaic_config = dict(enable_blocking=True, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-    # qkv blocking
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-    # head qkv blocking
-    qaic_config = dict(
-        enable_blocking=True,
-        head_block_size=HEAD_BLOCK_SIZE,
-        num_kv_blocks=NUM_KV_BLOCKS,
-        num_q_blocks=NUM_Q_BLOCKS,
-    )
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-
-@pytest.mark.dummy_layers
-@pytest.mark.llm_model
-@pytest.mark.on_qaic
-@pytest.mark.parametrize("model_name", test_models_blockedKV[:1])
-def test_dummy_causal_all_blocking_pytorch_vs_kv_vs_ort_vs_ai100_CB(model_name, manual_cleanup):
-    HEAD_BLOCK_SIZE = 8
-    NUM_KV_BLOCKS = 2
-    NUM_Q_BLOCKS = 2
-
-    hf_config = AutoConfig.from_pretrained(
-        model_name,
-        trust_remote_code=model_name in ModelConfig.EXTERNAL_MODELS,
-        **model_config_dict[model_name].get("additional_params", {}),
-    )
-    n_layer = -1
-    if model_name in ModelConfig.QUANTIZED_MODELS:
-        n_layer = get_custom_n_layers(model_name)
-        hf_config = None
-    # head blocking only
-    qaic_config = dict(enable_blocking=True, head_block_size=HEAD_BLOCK_SIZE)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        config=hf_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-    # kv blocking only
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        config=hf_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-    # q block only
-    qaic_config = dict(enable_blocking=True, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        config=hf_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-    # qkv blocking
-    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS, num_q_blocks=NUM_Q_BLOCKS)
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        config=hf_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
-
-    # head qkv blocking
-    qaic_config = dict(
-        enable_blocking=True,
-        head_block_size=HEAD_BLOCK_SIZE,
-        num_kv_blocks=NUM_KV_BLOCKS,
-        num_q_blocks=NUM_Q_BLOCKS,
-    )
-    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
-        model_name=model_name,
-        qaic_config=qaic_config,
-        n_layer=n_layer,
-        config=hf_config,
-        manual_cleanup=manual_cleanup,
-        continuous_batching=True,
-    )
